Updates to model and inclusion of README
Model now includes a basic implementation of the Keypad AO, and the beginnings of the Shelf AO. Some implementation of the internal data structures for the shelf has also been attempted, although this is likely to be subject to heavy redesign/refactor in the near future.
Some BSP implementation for the Keypad has also been done; a simple version for the Arduino port, and a corresponding version for the Win32 port - these will require some decent testing!